AntiDote-master/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.12.0
django-google-maps==0.12.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.2.0
joblib==1.0.1
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
